Port teehistorian to DDNet 0.7
This uses version 2 of the teehistorian file format which has support
for extra chunks, i.e. rarely used chunks that need not be standardized.

teehistorian records all network messages and inputs from the players as
well as the player positions in each tick. It stores this info in a
highly compressible output format.

It skips private information like chat messages and votes.

Also add a couple of tests.